requirements.txt
Django==2.2
djangorestframework==3.4.1
